Move ErrorCallout and SuccessCallout inside Widget folder
